making some addl changes for testing

+++ ass3/SafeDeleteSuite/trash.sh
@@ -1,6 +1,6 @@
 #!/bin/bash                                                                                       
 if [[ -d $TRASH ]]; then
-    /bin/rm -r "${TRASH}"/*
+    /bin/rm -r "${TRASH:?}"/*
     exit 0
 else
     echo "Could not find Trash. Check you trash directory to make sure that it is set up and that TRASH variable is set up"
